7.30 enums ✅ modify structs1 i16 -> u8

+++ exercises/structs/structs1.rs
@@ -8,9 +8,9 @@
 // 1. 正常结构体
 struct ColorClassicStruct {
     // TODO: Something goes here
-    red: i16,
-    green: i16,
-    blue: i16,
+    red: u8,
+    green: u8,
+    blue: u8,
 }
 
 // 2. tuple 结构体
